+++ yrt-pet/unit_tests/operators/test_Siddon_GPU.cu
@@ -23,9 +23,9 @@ TEST_CASE("siddon_gpu_vs_cpu", "[siddon-gpu]")
 	const size_t numDets = scanner->getNumDets();
 
 	// Setup image
-	constexpr int nx = 256;
-	constexpr int ny = 256;
-	constexpr int nz = 128;
+	constexpr ssize_t nx = 256;
+	constexpr ssize_t ny = 256;
+	constexpr ssize_t nz = 128;
 	constexpr float ox = 3.0f;
 	constexpr float oy = 10.0f;
 	constexpr float oz = -15.0f;
