+++ yrt-pet/unit_tests/test_utils.cpp
@@ -151,8 +151,8 @@ bool checkImageAllPositive(const Image& img)
 	const ImageParams& params = img.getParams();
 
 	const float* i_ptr = img.getRawPointer();
-	const int numVoxels = params.nx * params.ny * params.nz * params.nt;
-	for (int i = 0; i < numVoxels; i++)
+	const ssize_t numVoxels = params.nx * params.ny * params.nz * params.nt;
+	for (ssize_t i = 0; i < numVoxels; i++)
 	{
 		if (i_ptr[i] < 0)
 		{
